backend/requirements.txt
fastapi
uvicorn
# celery
# redis
openai-whisper
python-multipart
pydantic
sqlalchemy